refactored per https://github.com/pycaret/pycaret/issues/1748#issuecomment-1059967672
