+++ pycaret/internal/pycaret_experiment/__init__.py
@@ -1,20 +1,22 @@
-from .anomaly_experiment import AnomalyExperiment
-from .clustering_experiment import ClusteringExperiment
+# from .anomaly_experiment import AnomalyExperiment
+# from .clustering_experiment import ClusteringExperiment
 
-from .classification_experiment import ClassificationExperiment
-from .regression_experiment import RegressionExperiment
-from .time_series_experiment import TSForecastingExperiment
-from .time_series_experiment import TimeSeriesExperiment
+# from .classification_experiment import ClassificationExperiment
+# from .regression_experiment import RegressionExperiment
+# from pycaret.time_series import TSForecastingExperiment
 
-from .utils import MLUsecase
+# from .utils import MLUsecase
 
-
-def experiment_factory(usecase: MLUsecase):
-    switch = {
-        MLUsecase.CLASSIFICATION: ClassificationExperiment,
-        MLUsecase.REGRESSION: RegressionExperiment,
-        MLUsecase.CLUSTERING: ClusteringExperiment,
-        MLUsecase.ANOMALY: AnomalyExperiment,
-        MLUsecase.TIME_SERIES: TSForecastingExperiment,
-    }
-    return switch[usecase]()
+################################################################################
+#### NOTE: experiment_factory is not being used anymore but was causing     ####
+#### circular import issues. Hence removing.                                ####
+################################################################################
+# def experiment_factory(usecase: MLUsecase):
+#     switch = {
+#         MLUsecase.CLASSIFICATION: ClassificationExperiment,
+#         MLUsecase.REGRESSION: RegressionExperiment,
+#         MLUsecase.CLUSTERING: ClusteringExperiment,
+#         MLUsecase.ANOMALY: AnomalyExperiment,
+#         MLUsecase.TIME_SERIES: TSForecastingExperiment,
+#     }
+#     return switch[usecase]()
